refactor game logs package into separate components for better readability
