+++ rcon/cli.py
@@ -14,19 +14,20 @@ import rcon.expiring_vips.service
 import rcon.seed_vip.service
 import rcon.user_config
 import rcon.user_config.utils
-from rcon import auto_settings, broadcast, game_logs, routines
+from rcon import auto_settings, broadcast, routines
 from rcon.automods import automod
 from rcon.blacklist import BlacklistCommandHandler
 from rcon.cache_utils import RedisCached, get_redis_pool, invalidates
 from rcon.discord_chat import get_handler
-from rcon.game_logs import LogLoop, LogStream, load_generic_hooks
+from rcon.logs.loop import LogLoop, load_generic_hooks
+from rcon.logs.recorder import LogRecorder
+from rcon.logs.stream import LogStream
 from rcon.models import PlayerID, enter_session, install_unaccent
 from rcon.rcon import get_rcon
 from rcon.scoreboard import live_stats_loop
 from rcon.steam_utils import enrich_db_users
 from rcon.user_config.auto_settings import AutoSettingsConfig
 from rcon.user_config.log_stream import LogStreamUserConfig
-from rcon.user_config.seed_vip import SeedVIPUserConfig
 from rcon.user_config.webhooks import (
     BaseMentionWebhookUserConfig,
     BaseUserConfig,
@@ -129,7 +130,7 @@ def run_blacklists():
 @click.option("-t", "--frequency-min", default=5)
 @click.option("-n", "--now", is_flag=True)
 def run_log_recorder(frequency_min, now):
-    game_logs.LogRecorder(frequency_min, now).run()
+    LogRecorder(frequency_min, now).run()
 
 
 def init(force=False):
