+++ rcon/watchlist.py
@@ -3,7 +3,7 @@ from datetime import datetime
 from discord_webhook import DiscordEmbed
 
 from rcon.discord import get_prepared_discord_hooks
-from rcon.game_logs import on_connected
+from rcon.logs.loop import on_connected
 from rcon.hooks import inject_player_ids
 from rcon.models import WatchList, enter_session
 from rcon.player_history import _get_set_player, get_player
